fix: 테스트 오류 수정

+++ src/test/kotlin/com/brifo/server/notification/repository/NotificationContentQueryRepositoryTest.kt
@@ -161,7 +161,7 @@ class NotificationContentQueryRepositoryTest {
                 Decision.create(
                     briefing = briefings.first(),
                     direction = DecisionDirection.UP,
-                    confidenceLevel = 3.toShort(),
+                    confidenceLevel = 3,
                 ),
             )
         val dailyPrice =
